+++ crates/uzumaki/src/element.rs
@@ -1,7 +1,7 @@
 use crate::cursor::UzCursorIcon;
 use crate::input::InputState;
 use crate::interactivity::Interactivity;
-use crate::style::{Bounds, TextSelectable, TextStyle, UzStyle};
+use crate::style::{Bounds, TextSelectable, UzStyle};
 use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
 use vello::peniko::Blob;
@@ -181,15 +181,6 @@ pub struct TextSelectRun {
     pub entries: Vec<TextRunEntry>,
     pub flat_text: String,
     pub total_graphemes: usize,
-}
-
-#[derive(Clone, Debug)]
-pub struct NodeContext {
-    pub dom_id: UzNodeId,
-    pub text: Option<TextContent>,
-    pub text_style: TextStyle,
-    pub is_input: bool,
-    pub image: Option<ImageMeasureInfo>,
 }
 
 pub struct ElementNode {
